# apps/gkx-web-app/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/session-contents": {
      "post": {
        "operationId": "SessionContentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionContentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session content created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Session content created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SessionContentModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Crear relacion session-content",
        "tags": [
          "Session Contents"
        ]
      },
      "get": {
        "operationId": "SessionContentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Session contents retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Session contents retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SessionContentModel"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Listar session-contents",
        "tags": [
          "Session Contents"
        ]
      }
    },
    "/api/session-contents/{id}": {
      "get": {
        "operationId": "SessionContentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identificador de session-content",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session content retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Session content retrieved successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SessionContentModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Obtener session-content por id",
        "tags": [
          "Session Contents"
        ]
      },
      "patch": {
        "operationId": "SessionContentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identificador de session-content",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionContentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session content updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Session content updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SessionContentModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Actualizar session-content",
        "tags": [
          "Session Contents"
        ]
      },
      "delete": {
        "operationId": "SessionContentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identificador de session-content",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session content deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Session content deleted successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeleteDataModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Eliminar session-content",
        "tags": [
          "Session Contents"
        ]
      }
    },
...
    "/api/training-sessions/{sessionId}/contents/{sessionContentId}": {
      "patch": {
        "operationId": "SessionContentsNestedController_update",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Identificador de la sesion",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sessionContentId",
            "required": true,
            "in": "path",
            "description": "Identificador de session-content",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionContentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session content updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Session content updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SessionContentModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Actualizar contenido de una sesion",
        "tags": [
          "Session Contents"
        ]
      },
      "delete": {
        "operationId": "SessionContentsNestedController_remove",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Identificador de la sesion",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sessionContentId",
            "required": true,
            "in": "path",
            "description": "Identificador de session-content",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session content deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Session content deleted successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeleteDataModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Eliminar contenido de una sesion",
        "tags": [
          "Session Contents"
        ]
      }
    },
...
    "/api/attendance/session/{sessionId}": {
      "get": {
        "operationId": "AttendanceController_findBySession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Identificador de la sesion",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session attendance retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Session attendance retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AttendanceModel"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Listar asistencias por sesion",
        "tags": [
          "Attendance"
        ]
      }
    },
...
  },
  "info": {
    "title": "GKX API",
    "description": "Documentacion de la API GKX",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Ingresa el token JWT en formato Bearer"
      }
    },
    "schemas": {
...
      "DeleteDataModel": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "deleted"
        ]
      },
...
      "SessionContentModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "trainingContentId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "taskName": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "notes": {
            "type": "object",
            "nullable": true
          },
          "customDurationMinutes": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "sessionId",
          "trainingContentId",
          "taskName",
          "order",
          "notes",
          "customDurationMinutes",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateSessionContentDto": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "trainingContentId": {
            "type": "string",
            "format": "uuid"
          },
          "taskName": {
            "type": "string",
            "example": "Entrenamiento"
          },
          "order": {
            "type": "number",
            "example": 1
          },
          "notes": {
            "type": "string"
          },
          "customDurationMinutes": {
            "type": "number",
            "example": 20
          }
        },
        "required": [
          "tenantId",
          "sessionId",
          "taskName"
        ]
      },
      "UpdateSessionContentDto": {
        "type": "object",
        "properties": {}
      },
...
      "AttendanceModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "goalkeeperId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "comment": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "sessionId",
          "goalkeeperId",
          "status",
          "comment",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}